refactor: delete proto/tests/common and use registry/registry
